+++ newbank/server/CustomerID.java
@@ -7,17 +7,12 @@ public class CustomerID {
 	private String password;
 	private String IBAN;
 
-	public CustomerID(String name, String userName, String password, String iban) {
-		this.Name = name; this.userName = userName; this.password = password; this.IBAN = iban;
+	public CustomerID(String name, String userName, String password) {
+		this.Name = name; this.userName = userName; this.password = password;
 	}
 
 	public String getName() {
 		return Name;
-	}
-
-	public String getIBAN()
-	{
-		return IBAN;
 	}
 
 	public String getUserName() {
